requirements.txt
# Used for when working from a virtualenv.
# Use this file by running "$ pip install -r requirements.txt"
Flask>=0.10
flask-oidc
Flask-wtf
fedora-messaging
kitchen
python-fedora
SQLAlchemy
wtforms==2.2.1
arrow
blinker # required by Flask for the unit-tests (for signaling)
six
gunicorn
psycopg2
rsa==4.0  # last version that supports python2
fasjson_client
fedora_elections_messages
